move everything into shared?
